fix: rustfmt formatting and reorganize tests into ≤200-line files
