+++ crates/loopal-runtime/tests/agent_loop/mod.rs
@@ -61,6 +61,7 @@ mod preflight_test;
 mod record_message_test;
 mod retry_cancel_test;
 mod run_test;
+mod stream_truncation_edge_test;
 mod stream_truncation_test;
 mod tools_test;
 mod turn_completion_edge_test;
