+++ crates/loopal-runtime/tests/agent_loop/stream_truncation_test.rs
@@ -124,77 +124,3 @@ async fn test_eof_empty_stream_exits_without_continue() {
     assert_eq!(output.terminate_reason, TerminateReason::Goal);
     assert_eq!(runner.turn_count, 1);
 }
-
-/// Prior successful turn → second turn truncates → first turn's output preserved.
-#[tokio::test]
-async fn test_truncation_preserves_prior_turn_output() {
-    let tmp = std::env::temp_dir().join(format!("la_prior_{}.txt", std::process::id()));
-    std::fs::write(&tmp, "x").unwrap();
-    let calls = vec![
-        // First LLM: text + tool (normal)
-        vec![
-            Ok(StreamChunk::Text {
-                text: "Step one done.".into(),
-            }),
-            Ok(StreamChunk::ToolUse {
-                id: "tc-1".into(),
-                name: "Read".into(),
-                input: serde_json::json!({"file_path": tmp.to_str().unwrap()}),
-            }),
-            Ok(StreamChunk::Done {
-                stop_reason: StopReason::EndTurn,
-            }),
-        ],
-        // Second LLM: truncated (text only, no Done)
-        vec![Ok(StreamChunk::Text {
-            text: "Starting step two...".into(),
-        })],
-        // Third LLM (auto-continue after truncation): final
-        vec![
-            Ok(StreamChunk::Text {
-                text: "Step two complete.".into(),
-            }),
-            Ok(StreamChunk::Done {
-                stop_reason: StopReason::EndTurn,
-            }),
-        ],
-    ];
-    let (mut runner, mut event_rx) = make_multi_runner(calls);
-    tokio::spawn(async move { while event_rx.recv().await.is_some() {} });
-
-    let output = runner.run().await.unwrap();
-    assert_eq!(output.result, "Step two complete.");
-    assert_eq!(output.terminate_reason, TerminateReason::Goal);
-    let _ = std::fs::remove_file(&tmp);
-}
-
-/// Truncation with only tool_uses (no text) → tools discarded, auto-continue.
-#[tokio::test]
-async fn test_truncation_tool_only_discards_and_continues() {
-    let tmp = std::env::temp_dir().join(format!("la_toolonly_{}.txt", std::process::id()));
-    std::fs::write(&tmp, "y").unwrap();
-    let path_json = serde_json::json!({"file_path": tmp.to_str().unwrap()});
-    let calls = vec![
-        // First LLM: tool only, no Done (truncated)
-        vec![Ok(StreamChunk::ToolUse {
-            id: "tc-1".into(),
-            name: "Read".into(),
-            input: path_json.clone(),
-        })],
-        // Second LLM (auto-continue): tool + Done
-        vec![
-            Ok(StreamChunk::ToolUse { id: "tc-2".into(), name: "Read".into(), input: path_json }),
-            Ok(StreamChunk::Done { stop_reason: StopReason::EndTurn }),
-        ],
-        // Third LLM: final text
-        vec![
-            Ok(StreamChunk::Text { text: "Read complete.".into() }),
-            Ok(StreamChunk::Done { stop_reason: StopReason::EndTurn }),
-        ],
-    ];
-    let (mut runner, mut event_rx) = make_multi_runner(calls);
-    tokio::spawn(async move { while event_rx.recv().await.is_some() {} });
-    let output = runner.run().await.unwrap();
-    assert_eq!(output.result, "Read complete.");
-    let _ = std::fs::remove_file(&tmp);
-}
